Move binding of Populations from templates.h into populations.h
